extract from mappings stimulus the drawBubble logic in a file
